# Conflicts:
#	math_utils/linear_approx.h
#	tests/linear_approx_tests.c

+++ math_utils/lin_cos.c
@@ -1,12 +1,14 @@
 #include <stdlib.h>
 #include <math.h>
 #include <assert.h>
+#include <stdint.h>
 #include "linear_approx.h"
 
 #define Y_INT (1<<SF_ATAN_IN)    // magnitude of y-intercepts is 1, scaled
 
-int lin_cos(int theta) {
-    int result;    
+int16_t lin_cos(int16_t theta)
+{
+    int16_t result;    
     
     if (theta < 0) {
         result = SLOPE_SINCOS * theta + Y_INT;
